Added defaultPlayerHealth to clear up abstraction between minimumPlayerHealth and the actual health the player will spawn with.
Added magicsethealth as a way to set the players current health (not maximum health).

Added magicleveluphealth as a command, relies on external permissions added to a player such as magichealth.rankuphealth.40 (twenty hearts) to allow the player to set their health to whatever permission they have. This can also be run on players with the permission magichealth.sethealth.